Aadded methods to add and update objects in DB.

+++ app/src/main/java/com/example/taras/weather/API/OpenWeatherMap/Wind.java
@@ -11,7 +11,7 @@ public class Wind {
     private double speed;
     @SerializedName("deg")
     @Expose
-    private long deg;
+    private double deg;
 
     public double getSpeed() {
         return speed;
@@ -21,11 +21,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public long getDeg() {
+    public double getDeg() {
         return deg;
     }
 
-    public void setDeg(long deg) {
+    public void setDeg(double deg) {
         this.deg = deg;
     }
